requirements.txt
torch==1.4.0
numpy==1.18.1
matplotlib==3.1.3
tqdm==4.43.0
mxnet_cu90==1.2.1
scipy==1.4.1
bcolz==1.2.1
easydict==1.7
opencv_python==3.4.0.12
Pillow==7.0.0
scikit_learn==0.19.2
torchvision==0.2.1
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
